% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{get.dist.loci}
\alias{get.dist.loci}
\title{Calculates the distance of a marker from some objective "truth".}
\usage{
get.dist.loci(loci, estmap, realmap)
}
\arguments{
\item{loci}{Character string or number specifying the marker name.}

\item{estmap}{Data frame in which has a column called "names" containing marker 
names and a column called "position" containing marker positions.}

\item{realmap}{Data frame in which the first column contains marker names and 
the second column marker positions. Column names are not necessary.}
}
\value{
position in estmap-position in realmap
}
\description{
Calculates the distance of a marker from some objective "truth".
}
\details{
Both the first column of \code{realmap} and \code{estmap$name} must contain 
\code{markername}, but aside from this they do not have to have identical entries.
}
