% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{get.nearest.informative}
\alias{get.nearest.informative}
\title{For a given marker finds the nearest neighbours with LOD scores > 0.}
\usage{
get.nearest.informative(loci, lodmap)
}
\arguments{
\item{loci}{Scalar indicating a marker number}

\item{lodmap}{Symmetric matrix of pairwise LOD scores}
}
\value{
A vector of length 1 or 2 containing the rank of the nearest informative markers.
}
\description{
Finds the nearest neighbours of a marker with LOD scores > 0.
}
\details{
The columns and rows of the matrix should be in the order corresponding to 
the estimated map order. The function then returns the ranks of first markers 
to the left and right of the marker of interest with non-zero lod scores.
}
