% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{invert.map}
\alias{invert.map}
\title{Invert the order of locimap from an estimated map.}
\usage{
invert.map(locimap)
}
\arguments{
\item{locimap}{a data frame containing the markers names and positions}
}
\value{
The original data frame in inverted order with the distances inverted so that the end marker is now the starting marker.
}
\description{
Takes the locimap from estimate.maps a dataframe containing names and positions 
and any other information in increasing order of distance and inverts the order.
}
\details{
The map should be a data frame with a column called 'position'. It should have a starting marker a position zero. 
The function then inverts the distances from so that the marker at maximum distance from the starting marker (the end marker) is at distance 0 
and the original starting marker is now at the maximum distance. It also inverts the order of the rows in the data frame. 
Thus if the markers were originally in order of increasing distance from the starting marker they will now be in order of increasing distance 
from the end marker.
}
